Add Redis Cache

+++ score_sheet/Repository/Repo.cs
@@ -1,31 +1,54 @@
 ﻿using score_sheet.Model;
+using score_sheet.RedisCache;
 
 namespace score_sheet.Repository
 {
     public class Repo:IRepo
     {
         private readonly ScoreContext context;
-        public Repo(ScoreContext context)
+        private readonly IRedisCache redisCache;
+        public Repo(ScoreContext context, IRedisCache redisCache)
         {
             this.context=context;
+            this.redisCache=redisCache;
         }
 
         public void AddScore(ScoreSheet scoresheet)
         {
             context.Add(scoresheet);
             context.SaveChanges();
+            redisCache.RemoveData("AllScoreSheets");
         }
 
         public void DeleteScore(int RollNo)
         {
             var sc = GetScoreByRollNo(RollNo);
-            context.scoresheets.Remove(sc);
-            context.SaveChanges();
+            if (sc != null)
+            {
+                context.scoresheets.Remove(sc);
+                context.SaveChanges();
+
+                // When deleting a score, remove any cached data related to scoresheets
+                redisCache.RemoveData("AllScoreSheets");
+            }
         }
 
         public List<ScoreSheet> GetAll()
         {
-            return context.scoresheets.ToList();
+            // Check if data exists in cache
+            var cachedData = redisCache.GetData<List<ScoreSheet>>("AllScoreSheets");
+            if (cachedData != null)
+            {
+                return cachedData;
+            }
+
+            // If not cached, retrieve data from the database
+            var allScoresheets = context.scoresheets.ToList();
+
+            // Cache the data for future use
+            redisCache.SetData("AllScoreSheets", allScoresheets, DateTimeOffset.Now.AddDays(1));
+
+            return allScoresheets;
         }
 
         public ScoreSheet GetResult(ScoreSheet scoresheet)
@@ -36,17 +59,39 @@ namespace score_sheet.Repository
 
         public ScoreSheet GetScoreByRollNo(int RollNo)
         {
-            return context.scoresheets.Where(o => o.RollNo == RollNo).FirstOrDefault();
+            // Check if data exists in cache for the specific RollNo
+            var cachedScore = redisCache.GetData<ScoreSheet>($"ScoreByRollNo:{RollNo}");
+            if (cachedScore != null)
+            {
+                return cachedScore;
+            }
+
+            // If not cached, retrieve data from the database
+            var score = context.scoresheets.FirstOrDefault(o => o.RollNo == RollNo);
+
+            // Cache the data for future use
+            if (score != null)
+            {
+                redisCache.SetData($"ScoreByRollNo:{RollNo}", score, DateTimeOffset.Now.AddDays(1));
+            }
+
+            return score;
         }
 
         public void UpdateScore(int RollNo, ScoreSheet scoresheet)
         {
             var sc = GetScoreByRollNo(RollNo);
-            sc.Name = scoresheet.Name;
-            sc.Marks1 = scoresheet.Marks1;
-            sc.Marks2 = scoresheet.Marks2;
-            sc.Marks3 = scoresheet.Marks3;
-            context.SaveChanges();
+            if (sc != null)
+            {
+                sc.Name = scoresheet.Name;
+                sc.Marks1 = scoresheet.Marks1;
+                sc.Marks2 = scoresheet.Marks2;
+                sc.Marks3 = scoresheet.Marks3;
+                context.SaveChanges();
+
+                // When updating a score, remove any cached data related to scoresheets
+                redisCache.RemoveData("AllScoreSheets");
+            }
         }
     }
 }
